Added Various Elements and Polish
Added Rocks, changed Lighting and skybox to simulate the underworld aesthetic, mirrored the map to simulate the metaphorical symbolism from the tale of Sisyphus where once the player gets to the top the boulder falls down again. Changed the jump height of the player so they cant jump  out of bounds.